fixed bug that occurs when viewing a discussion
