+++ src/js/discussion/discuss/DiscussionList.js
@@ -3,6 +3,7 @@ import Discussion from './Discussion'
 import { connect } from 'react-redux'
 import { Col, Button, ButtonGroup } from 'react-bootstrap';
 import { Link } from 'react-router'
+import { getCamps } from '../camps/campActions'
 require('./styles.css')
 
 class DiscussionList extends React.Component{
@@ -13,12 +14,17 @@ class DiscussionList extends React.Component{
     }
   }
 
+  fetchCamps(discussionId, fullname) {
+    this.props.getCamps(discussionId, fullname)
+  }
+
   render() {
     return (
       <div className='discussionsList'>
         {Object.keys(this.props.discussionsList).length > 0 && Object.keys(this.props.discussionsList).map((discussionId, index) =>
-        <ButtonGroup vertical block className='buttonGroupEl'>
-          <div className='topic'></div>
+        <ButtonGroup vertical block className='buttonGroupEl'
+          onClick={()=> this.fetchCamps(discussionId, this.props.fullname)}
+        >
           <Link to={`/discuss/${discussionId}`}>
             <Button className="col-md-10 col-md-offset-1 discussButton">
               <Discussion
@@ -37,10 +43,19 @@ class DiscussionList extends React.Component{
 
 const mapStateToProps = (state) => {
   return {
+    fullname: state.profileReducer.fullname,
     discussionsList: state.discussionsGet.discussions
   }
 } //we need the state.discussions array to pass into the DiscussionList as props
 //we'd also need the articles retrieved from the get request
 //DiscussionList then displays the discussions //
 
-export default connect(mapStateToProps)(DiscussionList)
+const mapDispatchToProps = (dispatch) => {
+  return {
+    getCamps: (discussionId, fullname) => {
+      dispatch(getCamps(discussionId, fullname))
+    }
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(DiscussionList)
